Added Users and Laravel UI & Authentication

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Post;
 
@@ -12,7 +13,8 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::all();
+        // $posts = Post::all();
+        $posts = Auth::user()->posts;
         return view('Posts.index', compact('posts'));
     }
 
@@ -43,6 +45,9 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if (Auth::id() != $post->user_id) {
+            abort(403);
+        }
         return view('Posts.show', compact('post'));
     }
 
@@ -51,6 +56,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if (Auth::id() != $post->user_id) {
+            abort(403);
+        }
         return view('Posts.edit', compact('post'));
     }
 
@@ -59,6 +67,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if (Auth::id() != $post->user_id) {
+            abort(403);
+        }
         $request->validate([
             'title' => 'required',
             'content' => 'required'
@@ -74,6 +85,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::id() != $post->user_id) {
+            abort(403);
+        }
         $post->delete();
 
         return redirect()->route('posts.index');
